pyproject.toml
[project]
name = "discord-guild-configurator"
version = "0.1.0"
description = "Discord guild configuration manager"
readme = "README.md"
authors = [
    { name = "Niklas Mertsch", email = "niklasmertsch@web.de" },
]
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.6,<3",
    "pydantic>=2.12.5,<3",
]

[project.scripts]
discord-guild-configurator = "discord_guild_configurator.main:main"

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "prek>=0.3.3",
    "ruff>=0.15.2",
    "ty>=0.0.18",
]

[tool.ty.terminal]
output-format = "concise"

[tool.ty.rules]
all = "error"

[tool.ruff]
line-length = 100
unsafe-fixes = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    # incompatible with `ruff format`
    "COM812",
    # (f-)string in exception and logging
    "EM101",
    "EM102",
    "TRY003",
    "G004",
    # missing docstrings
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D107",
]
fixable = [
    "TC",  # type checking
    "ANN",  # type annotations
    "F401",  # unused imports
    "I001",  # unsorted imports
    "UP",  # pyupgrade
    "PLR1730",  # if min/max over if else
    "RET505",  # superfluous else-return
    "PLR5501",  # collapse else-if into elif
    "F541",  # superfluous f-string prefix
    "F841",  # unused variable
    "B007",  # unused loop variable
    "PIE790",  # unnecessary pass
    "Q003",  # change single/double quotes
]

[tool.ruff.lint.per-file-ignores]
"configs/*" = [
    "INP001",  # no `__init__.py`
]
"scripts/*" = [
    "INP001",  # no `__init__.py`
    "T201",  # print
]

[tool.ruff.lint.pydocstyle]
convention = "pep257"

[tool.ruff.lint.flake8-tidy-imports.banned-api]
"pydantic.BaseModel" = { msg = "Use discord_build_configurator._utils.StrictBaseModel instead" }

[tool.ruff.lint.flake8-type-checking]
runtime-evaluated-base-classes = ["discord_guild_configurator._utils.StrictBaseModel"]
